backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
passlib==1.7.4
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyMySQL==1.0.2
pyodbc==4.0.32
python-dotenv==0.19.2
python-jose==3.3.0
PyYAML==6.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
watchgod==0.7
websockets==10.1
psycopg2
